Move CSS injection logic to ThemableMixin package
