+++ packages/component-base/package.json
@@ -35,8 +35,7 @@
     "@polymer/polymer": "^3.0.0",
     "@vaadin/vaadin-development-mode-detector": "^2.0.0",
     "@vaadin/vaadin-usage-statistics": "^2.1.0",
-    "lit": "^3.0.0",
-    "style-observer": "^0.0.5"
+    "lit": "^3.0.0"
   },
   "devDependencies": {
     "@vaadin/chai-plugins": "24.8.0-alpha13",
