+++ packages/vaadin-themable-mixin/package.json
@@ -32,7 +32,8 @@
   ],
   "dependencies": {
     "@open-wc/dedupe-mixin": "^1.3.0",
-    "lit": "^3.0.0"
+    "lit": "^3.0.0",
+    "style-observer": "^0.0.5"
   },
   "devDependencies": {
     "@polymer/polymer": "^3.0.0",
